035:  - Refactored utilities out of index.js
